Modify to use BTRFS on root
This required a lot of modifications to the structure. Most importantly,
a larger portion of the setup is now done in `first-boot.sh` after the
first reboot because setting up snapshots didn't work until I rebooted
into the new system. I also took out some of the options to simplify
things.